serverless/agent-proxy/requirements.txt
azure-functions==1.21.3
azure-identity>=1.17.0
azure-ai-projects>=2.0.0b2
